+++ aquadoggo/src/graphql/types/mod.rs
@@ -3,9 +3,21 @@
 mod document;
 mod document_fields;
 mod document_meta;
+mod filter_input;
+mod filters;
 mod next_arguments;
+mod ordering;
+mod paginated_response;
 
-pub use document::Document;
+pub use document::{DocumentSchema, DocumentValue, PaginatedDocumentSchema};
 pub use document_fields::DocumentFields;
 pub use document_meta::DocumentMeta;
+pub use filter_input::{FilterInput, MetaFilterInput};
+pub use filters::{
+    BooleanFilter, DocumentIdFilter, DocumentViewIdFilter, FloatFilter, IntegerFilter, OwnerFilter,
+    PinnedRelationFilter, PinnedRelationListFilter, RelationFilter, RelationListFilter,
+    StringFilter,
+};
 pub use next_arguments::NextArguments;
+pub use ordering::{OrderBy, OrderDirection};
+pub use paginated_response::{PaginatedResponse, PaginationData};