+++ aquadoggo/src/lib.rs
@@ -32,7 +32,6 @@
 //!
 //! [`Tauri`]: https://tauri.studio
 #![warn(
-    missing_copy_implementations,
     missing_debug_implementations,
     missing_docs,
     trivial_casts,